move dmt5 video from top to side
